respond to kevin's feedback
