+++ packages/perseus/src/widgets/__testdata__/base-radio_testdata.js
@@ -1,0 +1,23 @@
+// @flow
+import type {ChoiceType} from "../radio/base-radio.jsx";
+
+export function generateChoice(options: $Partial<ChoiceType>): ChoiceType {
+    const base = {
+        checked: false,
+        crossedOut: false,
+        content: "",
+        rationale: "",
+        hasRationale: false,
+        showRationale: false,
+        showCorrectness: false,
+        correct: false,
+        originalIndex: 0,
+        isNoneOfTheAbove: false,
+        highlighted: false,
+        previouslyAnswered: false,
+        revealNoneOfTheAbove: false,
+        disabled: false,
+    };
+
+    return {...base, ...options};
+}
